+++ open-metadata-implementation/adapters/open-connectors/integration-connectors/csv-lineage-import-integration-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/csvlineageimporter/CSVLineageImporterConnector.java
@@ -128,7 +128,7 @@ public class CSVLineageImporterConnector extends LineageIntegratorConnector
 
             long numberOfRecords = csvFileStoreConnector.getRecordCount();
 
-            for (int i = 0; i < numberOfRecords; i++)
+            for (long i = 0; i < numberOfRecords; i++)
             {
                 List<String> columns = csvFileStoreConnector.readRecord(i);
